New SQLRPGLE program for our integration project.
This program first checks for available LOT inventory and returns the available quantity or requested quantity whichever is less.
if no inventory is found it then does a calculation to determine if we can produce the requested amount by the time of the requested ship date.  If so, the requested amount is returned as the quantity available.